feat: adding apiroutes, singleton, factorymethods and abstractFactory

+++ internal/domain/dtos/request_information_metadata.dto.go
@@ -1,6 +1,6 @@
 package dtos
 
 type RequestInformationMetadata struct {
-	Limit  int
-	Offset int
+	Limit  int64
+	Offset int64
 }
